% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state return a point graphic representing the distribution
of accidents for a given state number and for a given year. This
function calls make_filename, fars_read.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Interger which represents a state}

\item{year}{Integer which represents a target year}
}
\value{
If the inputs are correct, it returns graphic points where
graphic represents the state and each point represent a accident.
If state number is unknown in the file, it returns a error like
"invalid STATE number : {state.num}". If there is no accident this
year in this state, return "no accidents to plot" and NULL value
}
\description{
fars_map_state return a point graphic representing the distribution
of accidents for a given state number and for a given year. This
function calls make_filename, fars_read.
}
\examples{
fars_map_state(1, 2013)

}
