% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of characters referencing the path and name of
the CSV file to be processed}
}
\value{
This function returns a dplyr dataframe if the file exists.
Otherwize, it returns an error like "file '\code{filename}' doesn't
exist".
}
\description{
By default, this function reads and converts a CSV file into a dataframe.
It takes the CSV file as the only argument and it doesn't show the reading
progress bar.
}
