% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Groups together in a new dataframe all accidents by
selecting only the month and the year referenced in the archive files
"accident_{year}.csv.bz2" from the target years given in the input.
This function calls make_file_name and fars_read.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector with all the years to extract}
}
\value{
If the input is an integer vector as required, it returns
a dataframe with n+1 columns : MONTH and n years. Otherwise, when there
are no archive files with the target year, it returns a warning
like "invalid year: {year}" and NULL value.
}
\description{
Groups together in a new dataframe all accidents by
selecting only the month and the year referenced in the archive files
"accident_{year}.csv.bz2" from the target years given in the input.
This function calls make_file_name and fars_read.
}
\examples{
fars_read_years(c("2013", "2014")) -> success
fars_read_years(c(2013, 2014)) -> success
fars_read_years() -> error argument "years" is missing
fars_read_years(c(2013, 1991)) -> Warning :
not file associated to 1991, return NULL + df with only 2013

}
