% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years counts for each year given in input,
the number of accidents per MONTH. This function calls
fars_read_years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector with all the target years}
}
\value{
A dataframe with n+1 column : MONTH and n year.
Values of n year is the accident number for specific a month.
If the year doesn't exist in the files, return warning
}
\description{
fars_summarize_years counts for each year given in input,
the number of accidents per MONTH. This function calls
fars_read_years.
}
\examples{
fars_summarize_years(c(2014, 2015)) -> success

}
