% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename creates a dynamic filename with a year given in input.
The filename appear like "accident_{year}.csv.bz2". This function is
call in "fars_read_years".}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer which represents the desired year}
}
\value{
If the input type is a integer or can be convert into it, this
function return the filename "accident_{year}.csv.bz2".Else, returns
(1) the file "accident_NA.csv.bz2" and (2) a warning message
(NAs introduced by coercion)
}
\description{
make_filename creates a dynamic filename with a year given in input.
The filename appear like "accident_{year}.csv.bz2". This function is
call in "fars_read_years".
}
\examples{
make_filename(2013)
make_filename("2015")

}
